tidy up results page
